Rename blackboard in all files
