+++ bitbots_behavior/bitbots_body_behavior/bitbots_body_behavior/behavior_dsd/actions/deactivate_hcm.py
@@ -1,4 +1,4 @@
-from bitbots_blackboard.blackboard import BodyBlackboard
+from bitbots_blackboard.body_blackboard import BodyBlackboard
 from dynamic_stack_decider.abstract_action_element import AbstractActionElement
 from std_msgs.msg import Bool
