+++ bitbots_behavior/bitbots_body_behavior/bitbots_body_behavior/behavior_dsd/actions/speak.py
@@ -1,7 +1,7 @@
 import os
 
 from ament_index_python import get_package_share_directory
-from bitbots_blackboard.blackboard import BodyBlackboard
+from bitbots_blackboard.body_blackboard import BodyBlackboard
 from bitbots_tts.tts import speak
 from dynamic_stack_decider.abstract_action_element import AbstractActionElement
