+++ bitbots_behavior/bitbots_body_behavior/bitbots_body_behavior/behavior_dsd/decisions/pass_started.py
@@ -1,4 +1,4 @@
-from bitbots_blackboard.blackboard import BodyBlackboard
+from bitbots_blackboard.body_blackboard import BodyBlackboard
 from dynamic_stack_decider.abstract_decision_element import AbstractDecisionElement
